Add files via upload
The following is the data from a bank that has seen a drop in the verification of customers. The aim of the study is to know the key factor affecting the pass rate of the verification.
It can be seen that the overall pass rate is getting affected by the documentation verification step as shown in the plot.